Add vendor dependencies
- Ran dep prune to reduce size of vendor directory from ~270M to 23M
- Manually set dependency revision for github.com/manifoldco/promptui to 4baa118

Our source code depends on the Prompt stdin field to be of type io.ReadCloser. Unfortunately
the current latest release of promptui (v0.2.1) does not have this, so we have to set the
dependency to the latest commit hash on the master branch. This should be changed to a version
once promptui makes a new release.